requirement.txt
pyyaml
matplotlib
opencv-python
pillow<7
tqdm
numba
grpcio
grpcio-tools